Update 5.4.2.cpp
In the Zybook problem, with the original do while loop on this .cpp. It bring back this error. 
"Exited with return code -6 (SIGABRT).
terminate called after throwing an instance of 'std::out_of_range'
  what():  vector::_M_range_check: __n (which is 4) >= this->size() (which is 4)"

With my for loop if statement solution. It works without having to have the do while statement.